requirements.txt
imageio==2.26
joblib==1.1.0
jupyter==1.0.0
kornia==0.6.4
lpips==0.1.4
matplotlib==3.5.1
opencv-python==4.5.5.64
pandas==1.3.5
PyMCubes==0.1.2
PyYAML==6.0
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.7.3
trimesh==3.10.8
tensorboard
tqdm
yacs
partio
zstandard
msgpack
msgpack-numpy
plyfile
open3d
geomloss
